pyproject.toml
[tool.poetry]
name = "md-clean-up"
version = "0.1.0"
description = "A utility to create appropriate named links within markdown documents."
authors = ["Brian Arbuckle <brian@brianarbuckle.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
